voice-call-agent/livekit_agent/requirements.txt
# Minimal dependencies for LiveKit Cloud agent detection/build.
livekit-agents>=1.2.0
